requirements.txt
flask==3.0.0
flask-cors==4.0.0
supabase==2.9.0
requests==2.31.0
python-dotenv==1.0.0
websockets==13.1
realtime==2.0.5